aIIure/requirements.txt
pytest==2.9.0
allure-pytest==2.3.1b1
selenium>3.0.2
webium==1.2.1